chore: update readme
Added background image problem and the rationale behind the solution. In
order to manipulate the background image in the way it will need to be,
it has been added as a HTML element that can be rotated. This element is
removed from the accessibility tree, as it is purely cosmetic.